create a shared SubmitStepContent component
